Rilis v2309.0.0

+++ database/migrations/2022_12_08_023053_update_htaccess.php
@@ -6,6 +6,8 @@ use Illuminate\Database\Migrations\Migration;
 
 class UpdateHtaccess extends Migration
 {
+    public $root_default = '/var/www/html';
+
     /**
      * Run the migrations.
      *
@@ -14,38 +16,51 @@ class UpdateHtaccess extends Migration
     public function up()
     {
         $root_folder = env('ROOT_OPENSID');
-        // folder multisite
         $multisite = env('MULTISITE_OPENSID');
         $dirs = File::directories($multisite);
         $files = new Filesystem();
-        foreach ($dirs as $dir) {
-            $htopensid = $dir . DIRECTORY_SEPARATOR . '.htaccess';
-            $htapi = $dir . DIRECTORY_SEPARATOR . 'api-app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess';
-            $htpbb = $dir . DIRECTORY_SEPARATOR . 'pbb-app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess';
 
-            if (file_exists($htopensid)) {
-                File::delete($htopensid); // hapus .httacess opensid
-            } else {
-                exec('sudo ln -s ' . $root_folder . 'master-template' . DIRECTORY_SEPARATOR . 'template-desa' . DIRECTORY_SEPARATOR . '.htaccess ' . $htopensid);
+        if (file_exists($multisite)) {
+            foreach ($dirs as $dir) {
+                $master_template = $root_folder . 'master-template' . DIRECTORY_SEPARATOR;
+
+                if (file_exists($master_template)) {
+                    $this->updateHtaccess($root_folder, $master_template, $dir);
+                }
+
+                // perbarui index.php opensid
+                $file_index = $dir . DIRECTORY_SEPARATOR . 'index.php';
+                if (file_exists($file_index)) {
+                    $index = file_get_contents($file_index);
+                    $content = str_replace("symlink(SYMLINK_DOMAIN . DIRECTORY_SEPARATOR . 'pbb-app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess', '.htaccess')", "symlink(OPENSID_FOLDER . DIRECTORY_SEPARATOR . '.htaccess', '.htaccess')", $index);
+                    file_put_contents($file_index, $content);
+                }
             }
+        }
+    }
 
-            if (file_exists($htapi)) {
-                File::delete($htapi); // hapus .httacess api
-            } else {
-                exec('sudo ln -s ' . $root_folder . 'master-template' . DIRECTORY_SEPARATOR . 'template-api' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess ' . $htapi);
-            }
+    public function updateHtaccess($root_folder, $master_template, $dir)
+    {
+        $htopensid = $dir . DIRECTORY_SEPARATOR . '.htaccess';
+        $htapi = $dir . DIRECTORY_SEPARATOR . 'api-app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess';
+        $htpbb = $dir . DIRECTORY_SEPARATOR . 'pbb-app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess';
 
-            if (file_exists($htpbb)) {
-                File::delete($htpbb); // hapus .httacess pbb
-            } else {
-                exec('sudo ln -s ' . $root_folder . 'master-template' . DIRECTORY_SEPARATOR . 'template-pbb' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess ' . $htpbb);
-            }
+        if (file_exists($htopensid)) {
+            File::delete($htopensid); // hapus .httacess opensid
+        } else {
+            $root_folder == $this->root_default ? exec('sudo ln -s ' . $master_template . 'template-desa' . DIRECTORY_SEPARATOR . '.htaccess ' . $htopensid) : '';
+        }
 
-            // perbarui index.php opensid
-            $file_index = $dir . DIRECTORY_SEPARATOR . 'index.php';
-            $index = file_get_contents($file_index);
-            $content = str_replace("symlink(SYMLINK_DOMAIN . DIRECTORY_SEPARATOR . 'pbb-app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess', '.htaccess')", "symlink(OPENSID_FOLDER . DIRECTORY_SEPARATOR . '.htaccess', '.htaccess')", $index);
-            file_put_contents($file_index, $content);
+        if (file_exists($htapi)) {
+            File::delete($htapi); // hapus .httacess api
+        } else {
+            $root_folder == $this->root_default ? exec('sudo ln -s ' . $master_template . 'template-api' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess ' . $htapi) : '';
+        }
+
+        if (file_exists($htpbb)) {
+            File::delete($htpbb); // hapus .httacess pbb
+        } else {
+            $root_folder == $this->root_default ? exec('sudo ln -s ' . $master_template . 'template-pbb' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . '.htaccess ' . $htpbb) : '';
         }
     }
